requirements.txt
flake8==3.8.3
Flask==1.1.2
marshmallow==3.6.1
mongoengine==0.20.0
mongomock==3.19.0
python-dotenv==0.13.0
Werkzeug==1.0.1
pytest==5.4.3
gunicorn==19.9.0
Flask-Cors==3.0.9